Menu improvements
Fixed a couple of menu bugs and added controls for all the game parameters.
Also added a Memory Check Proof sequence of bytes so that the Arduino knows if the parameters have been successfully written at least once (so first-time run has defaults).  
This means that settings from previous installs will be overwritten by this code.